agent/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
openai==1.56.2
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
